+++ media/main.js
@@ -66,9 +66,9 @@ function handleMessage(message) {
       webviewDebugLogger(`Received prs: ${JSON.stringify(message.data)}`);
       addPRsToDoc(message.data.sortedPRs);
       // jira
-      if (message.data?.mostRelevantJiraTicket) {
+      if (message.data?.mostRelevantJiraTickets) {
         $("#mostRelevantJiraTicketHolder").empty();
-        addMostRelevantJiraTicket(message.data.mostRelevantJiraTicket);
+        addMostRelevantJiraTicket(message.data.mostRelevantJiraTickets);
       }
       clampCodeBlocks();
       break;