+++ src/extension.ts
@@ -224,15 +224,14 @@ export async function activate(context: vscode.ExtensionContext) {
         const parsedMessage = parsedCommitObject.message;
 
         // Jira
-        const mostRelevantJiraTicket =
+        const mostRelevantJiraTickets =
           (await getMostRelevantJiraTicket({
             user: session.account.label,
             prTitle: sortedPRs[0].title || parsedMessage,
           })) || {};
-
         provider.sendMessage({
           command: "prs",
-          data: { sortedPRs, uniqueBlames, mostRelevantJiraTicket },
+          data: { sortedPRs, uniqueBlames, mostRelevantJiraTickets },
         });
       } else {
         vscode.commands.executeCommand("watermelon.multiSelect");